existing skatches of the SBL objects
The library was growing over the last few years working with service broker. The code in this commit can be made to work, but it is not straight forward, self explanatory, or redy for generic use. The intention is to start here and gradually shape it into a truelly reusable tool.